feat: move stale dispatch reaper to fleet._dispatch_reaper, wire into boot handlers
- Create src/autoskillit/fleet/_dispatch_reaper.py (IL-2) with
  reap_stale_dispatches() and reap_stale_dispatches_async() extracted
  from cli/fleet/_fleet_lifecycle.py
- Add narrow IL-009 exception in pyproject.toml for the reaper's
  execution import (kill_process_tree, read_boot_id, read_starttime_ticks)
- Export reap_stale_dispatches and reap_stale_dispatches_async from
  autoskillit.fleet.__init__
- Replace _reap_stale_dispatches in cli/fleet/_fleet_lifecycle.py with
  a thin delegation wrapper; remove dead imports (psutil, time, TYPE_CHECKING)
- Wire reap_stale_dispatches_async into both _fleet_auto_gate_boot and
  _food_truck_auto_gate_boot in server/_lifespan.py, called before
  sweep_stale_dispatch_labels
- Update test_ast_rules.py allowed_files to include fleet/_dispatch_reaper.py
  and remove cli/fleet entries that no longer call kill_process_tree directly
- Simplify tests/cli/test_reap.py to delegation-only test
- Add importability guard to tests/cli/test_fleet_split.py
- Update fleet/CLAUDE.md and cli/fleet/CLAUDE.md with new file descriptions
- Update tests/fleet/CLAUDE.md with test_dispatch_reaper.py entry
