+++ src/autoskillit/fleet/__init__.py
@@ -8,6 +8,7 @@ from ._api import _write_pid as _write_pid
 from ._api import execute_dispatch
 from ._capture import CaptureCompletenessError
 from ._checkpoint_bridge import checkpoint_from_sidecar
+from ._dispatch_reaper import reap_stale_dispatches, reap_stale_dispatches_async
 from ._expressions import evaluate_skip_when
 from ._label_cleanup import (
     cleanup_orphaned_labels,
@@ -145,6 +146,8 @@ __all__ = [
     "find_dispatch_for_issue",
     "checkpoint_from_sidecar",
     "is_dispatch_session_alive",
+    "reap_stale_dispatches",
+    "reap_stale_dispatches_async",
     "FleetSemaphore",
     "IssueSidecarEntry",
     "append_sidecar_entry",
